# docs/swagger.yaml
definitions:
...
  internal_service.LoginRequest:
    properties:
      device_id:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  internal_service.MFAVerifyRequest:
    properties:
      code:
        type: string
      device_id:
        type: string
    type: object
...
  internal_service.SessionInfo:
    properties:
      created_at:
        type: string
      device_id:
        type: string
      device_info:
        type: string
      expires_at:
        type: string
      family_id:
        type: string
      ip_address:
        type: string
    type: object
...
host: localhost:3333
schemes:
  - http
info:
  contact:
    name: GoNet Drive Support
    url: https://github.com
  description: GoNet Drive is a self-hosted cloud file storage and sharing service.
  termsOfService: https://example.com/terms
  title: GoNet Drive API
  version: "1.0"
paths:
...
  /api/login:
    post:
      consumes:
      - application/json
      description: Authenticate with username and password. Returns JWT cookies. If
        MFA is enabled, a pre-auth token is returned and the client must call /api/mfa/verify.
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_service.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Login
      tags:
      - Auth
...
  /api/mfa/verify:
    post:
      consumes:
      - application/json
      description: Verify the TOTP code after receiving a pre-auth token from /api/login.
        On success, full auth cookies are set.
      parameters:
      - description: MFA verification
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_service.MFAVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Verify MFA Code
      tags:
      - Auth
  /api/refresh:
    post:
      description: Exchange a valid refresh token cookie for a new access/refresh
        token pair.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Refresh Token
      tags:
      - Auth
...
  /api/user/me:
    get:
      description: Get information about the currently authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Current User Info
      tags:
      - User
  /api/user/me/sessions:
    get:
      description: List all active refresh token sessions for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_service.SessionInfo'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Get Active Sessions
      tags:
      - User
  /api/user/me/sessions/{family_id}:
    delete:
      consumes:
      - application/json
      description: Revoke a specific refresh token family. Requires password confirmation.
      parameters:
      - description: Session family ID
        in: path
        name: family_id
        required: true
        type: string
      - description: Password confirmation
        in: body
        name: body
        required: true
        schema:
          properties:
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Revoke Session
      tags:
      - User
  /api/user/mfa/enable:
    post:
      consumes:
      - application/json
      description: Enable MFA after setting up a TOTP secret via /api/user/mfa/setup.
      parameters:
      - description: TOTP code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_service.MFAVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Enable MFA
      tags:
      - User
  /api/user/mfa/setup:
    get:
      description: Generate a TOTP secret and return the secret key and provisioning
        URL. Requires authentication.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Setup MFA
      tags:
      - User
...
  /api/user/status:
    get:
      description: Check if the current request is properly authenticated.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Auth Status Check
      tags:
      - User
...
  /api/user/ws:
    get:
      description: Establish a WebSocket connection for real-time file operation progress
        updates.
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: WebSocket Connection
      tags:
      - WebSocket
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT access token.
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    description: JWT access token stored as a cookie.
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"